% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence.Filter.OTUs.R
\name{Confidence.Filter.OTUs}
\alias{Confidence.Filter.OTUs}
\title{Filter Features (OTUS/SVs) such that they are present in at least X samples with at least a total of Y reads}
\usage{
Confidence.Filter.OTUs(OTUTABLE, MINSAMPS, MINREADS)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{MINSAMPS}{A minimum number of samples for the feature/OTU/SV to be observed in}

\item{MINREADS}{A minimum number of reads for a feature/OTUs/SV to be kept}
}
\value{
filtered.OTUTABLE
}
\description{
\code{Confidence.Filter.OTUs} Takes a table of FeaturesOTUs/SVs and filters for such that they are present in at least X samples with at least a total of Y reads
}

