% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence.Filter.R
\name{Confidence.Filter}
\alias{Confidence.Filter}
\title{Filter features such that they are present in at least X samples with at least a total of Y reads}
\usage{
Confidence.Filter(FEATURES, MINSAMPS, MINREADS, VERBOSE)
}
\arguments{
\item{FEATURES}{Table of feature/OTU/SV counts where Samples are columns, and feature IDs are row names}

\item{MINSAMPS}{A minimum number of samples for the feature/OTU/SV to be observed in}

\item{MINREADS}{A minimum number of reads for a feature/OTUs/SV to be kept across all samples}

\item{VERBOSE}{Should summary be printed? (T/F)}
}
\value{
filtered feature table
}
\description{
Takes a table of features and filters for such that they are present in at least X samples with at least a total of Y reads
}
