% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.OTUs.Fraction.R
\name{Filter.OTUs.Fraction}
\alias{Filter.OTUs.Fraction}
\title{\code{Filter.OTUs.Fraction} Remove Features/OTUs/SVs representing less than a certain specified fraction.}
\usage{
Filter.OTUs.Fraction(OTUTABLE, MINFRACTION)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names.}

\item{MINFRACTION}{A fraction to use to filter.}
}
\value{
Filtered Table
}
\description{
\code{Filter.OTUs.Fraction} Remove Features/OTUs/SVs representing less than a certain specified fraction.
}
