% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fraction.Filter.R
\name{Fraction.Filter}
\alias{Fraction.Filter}
\title{Filter.OTUs.Fraction}
\usage{
Fraction.Filter(FEATURES, MINFRACTION, VERBOSE)
}
\arguments{
\item{FEATURES}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names.}

\item{MINFRACTION}{A fraction to use to filter.}

\item{VERBOSE}{Should summary be printed? (T/F)}
}
\value{
Filtered Table
}
\description{
Remove Features/OTUs/SVs representing less than a certain specified fraction of the total dataset. Similar to filter_otus_from_otutable.py in QIIME.
}
