% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.Test.Data.R
\name{Generate.Test.Data}
\alias{Generate.Test.Data}
\title{\code{Generate.Test.Data} Generate random OTU table and metadata for testing purposes.}
\usage{
Generate.Test.Data(NSAMP, NFEATURE)
}
\value{
Returns list of mock OTU table [[1]] and metadata [[2]]
}
\description{
Creates a mock dataset based on a user defined number of samples and matching metadata for testing purposes. Returns a categorical variable called TestGroup and a continuous number called TestContinuous.
}
