% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.MicrobeR.Libs.R
\name{Load.MicrobeR.Libs}
\alias{Load.MicrobeR.Libs}
\title{\code{Load.MicrobeR.Libs} Import handy libraries}
\usage{
Load.MicrobeR.Libs()
}
\description{
dada2
ggplot2
reshape2
vegan
ZIBR
ALDEx2
seqinr
phangorn
gplots
phyloseq
GUniFrac
grid
doBy
RColorBrewer
randomForest
ggord
plyr
dplyr
lme4
lmerTest
multcomp
DT
}

