% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.CLR.R
\name{Make.CLR}
\alias{Make.CLR}
\title{\code{Make.CLR} Make a CLR-normalized Feature/OTU/SV table}
\usage{
Make.CLR(TABLE)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}
}
\value{
Table of CLR-normalized Abundances
}
\description{
Takes a table of Features/OTUs/SVs and converts centered log ratio abundances. See:
 Compositional analysis: a valid approach to analyze microbiome high-throughput sequencing data
  Gregory B. Gloor, Gregor Reid; Canadian Journal of Microbiology, 2016, 62:692-703, 10.1139/cjm-2015-0821
}

