% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.Proportion.R
\name{Make.Proportion}
\alias{Make.Proportion}
\title{\code{Make.Proportion} Takes a table of Features/OTUs/SVs and converts to Proportion (Sum to 1).}
\usage{
ProportionTab<-Make.Proportion(FeatureTable)
}
\arguments{
\item{FEATURES}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}
}
\value{
Table of Proportions
}
\description{
\code{Make.Proportion} Takes a table of Features/OTUs/SVs and converts to Proportion (Sum to 1).
}
