% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.Tree.R
\name{Make.Tree.QIIME}
\alias{Make.Tree.QIIME}
\title{\code{Make.Tree.X} Generate a phylogenetic tree and optional plot}
\usage{
Make.Tree.QIIME(FEATURENAMES, FEATURESEQS, MUSCLE, MAKEPHY, PRINTTREE)
}
\arguments{
\item{FEATURENAMES}{The names of your features/OTUs/SVs. Will be the row names of your input table in most cases.}

\item{FEATURESEQS}{The corresponding sequences for each feature. If this is a dada2 table, then feature names are also feature seqs.}

\item{MUSCLE}{(optional) The location of muscle, defaults to /macqiime/bin/muscle}

\item{MAKEPHY}{(optional) Location of qiime make_phylogeny.py, defatuls to /macqiime/anaconda/bin/make_phylogeny.py}

\item{PRINTTREE}{(optional) Would you like a tree to be printed TRUE/FALSE? defaults to TRUE}
}
\value{
Returns tree object and prints a phylogenetic tree
}
\description{
Takes a series of sequences and uses muscle and make_phylogeny (Make.Tree.QIIME) OR DECIPHER and phangorn (Make.Tree.R) as per https://f1000research.com/articles/5-1492/v2. The R-method will take considerably longer and becomes unusable for >a few hundred features. For QIIME method: can be used with MacQIIME or QIIME. Will default to MacQIIME locations. A file called Feature_Seqs.tree will be created in your working directory for future use.  Uses midpoint rooting, and fasttree. NOTE: This is using system commands... probabily will only work on mac and linux. If you are using windows reconsider your life choices and consider dual booting Ubuntu.
}
