% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.Heatmap.R
\name{Microbiome.Heatmap}
\alias{Microbiome.Heatmap}
\title{\code{Microbiome.Heatmap.R} Plot Heat map of microbiome data}
\usage{
Microbiome.Heatmap(OTUTABLE, METADATA, NTOPFEATURES, TRANSFORM, CATEGORY, ORDER,
  USERORDER, PLOTMETHOD)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{METADATA}{Metadata file to be used for blocking.}

\item{NTOPFEATURES}{The N most abundance features to plot (defaults to all). Calculated by taking largest row sums of percentage data}

\item{TRANSFORM}{Method to transform data with for plotting, valid options are log2, log10, clr, percent, zscore or none (defaults to log10). None would be ideal if for example a list of fold changes was supplied. Zscore is calculated on clr.}

\item{CATEGORY}{(optional) Category to create separate blocks (optional, defaults to order of samples in otutable)}

\item{ORDER}{(optional) An order of levels in Category to use, (optional, defaults to order of unique(Category)). Only works with heatmap.2 method.}

\item{USERORDER}{(optional) A vector with user specified sample order for plot with NA for separators, overides metadata, category and order. Only works with heatmap.2 method.}

\item{PLOTMETHOD}{(optional) Can be plotted using heatmap.2 or ggplot2, defaults to ggplot2}
}
\value{
Prints a heat map
}
\description{
Creates a heat map based based on user provided table. Transforms to log2(percent), log10(percent), or CLR as requested. For plotting purposes, a prior of 0.01% is added to the log percent abundances. Row clustering of ggplot2 output is by upgma of euclidean dist. If no metadata passed, column clustering will be applied in the heatmap.2 output.
}

