% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nice.Table.R
\name{Nice.Table}
\alias{Nice.Table}
\title{\code{Nice.Table} Print interactive table for markdown documents using datatable from DT with the options for export preincluded.}
\usage{
Nice.Table(TABLE, NROW)
}
\arguments{
\item{TABLE}{Any table to be printed}

\item{NROW}{(optional) number of rows to display per page, defaults to 10}
}
\description{
Function for the following command DT::datatable(TABLE, extensions='Buttons', options=list(pageLength=LENGTH, dom='bfrtip', buttons=c('copy','csv','excel', 'pdf') ))
}

