% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCoA.R
\name{PCoA}
\alias{PCoA}
\title{\code{PCoA} Plot PCoAs of Bray-Curtis and (Un)Weighted UniFrac using first 3 axes.}
\usage{
PCoA("Treatment_and_Time.pdf", experiment_metadata, otutable, ggtree, c("Treatment","Time","ExtractionPlate"), FALSE)
}
\arguments{
\item{PDFTITLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{METADATA}{Metadata Table with variables to color PCoAs by}

\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{TREE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{CONDS}{A vector of column names from METADATA to use: ex c("Treatment","Time")}

\item{SUBSAMPLE}{Should table be subsampled? TRUE/FALSE (default: TRUE)}
}
\description{
Will create a 2x6 PDF for each condition passed. UniFrac is implimented as per GUniFrac and PCoA from APE. Plotting via ggplot2.
}

