% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCoA3D.R
\name{PCoA3D}
\alias{PCoA3D}
\title{\code{PCoA3D} Generate Distance/Dissimilarities and plot PCoA in interactive 3D plot}
\usage{
PCoA(METRIC="braycurtis", METADATA=experiment_metadata, FEATURES=otutable, TREE=ggtree, COLOR="Time", SHAPE="Group", SUBSAMPLE=TRUE, AXIS=c(1,2,3), PALETTE=c("red","green"))
}
\arguments{
\item{METRIC}{Desired beta-diversity metric, options are Bray Curtis (braycurtis), Weighted UniFrac (weightedunifrac), UnWeighted UniFrac (unweightedunifrac), Jensen-Shannon diversgence (jsd).}

\item{METADATA}{Metadata table with variables to color or shape points by where sample is row name.}

\item{FEATURES}{Table of feature counts where Samples are columns, and IDs are row names.}

\item{TREE}{A phylogenetic tree of phylo class *only required for UniFrac Metrics}

\item{COLOR}{Metadata column to color samples by. Defaults to none.}

\item{SHAPE}{Metadata column to shape samples by. Defaults to none.}

\item{SUBSAMPLE}{Should table be subsampled? TRUE/FALSE (default: TRUE)}

\item{AXIS}{Which PCs should be plotted? Expects a numeric vector of length 3. defaults to  AXIS=c(1,2,3)}

\item{PALETTE}{a vector of COLORS to be used? defaults to the rainbow palette.}
}
\value{
Returns plot.
}
\description{
UniFrac is implimented as per Phyloseq, Bray Curtis from Vegan, Jensen-Shannon divergence from Phyloseq, and PCoA from APE. Plotting via plotly's 3d plotting.
}
