% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.Filter.R
\name{Read.Filter}
\alias{Read.Filter}
\title{\code{Read.Filter} Remove samples with less than a user specified number of reads.}
\usage{
Read.Filter(OTUTABLE, METADATA, READCUTOFF, VERBOSE, PLOT)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{METADATA}{Metadata file where samples are row names.}

\item{READCUTOFF}{Minimum number of reads to pass filter}

\item{VERBOSE}{Should lists of which samples were removed or kept be printed to screen? (defaults to TRUE)}

\item{PLOT}{Should a plot be made? (defaults to TRUE)}
}
\value{
A list containing new table of features[[1]] (Ex. OTU table) and Metadata[[2]].
}
\description{
\code{Read.Filter} Remove samples with less than a user specified number of reads.
}
