% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.Taxa.R
\name{Summarize.Taxa}
\alias{Summarize.Taxa}
\title{Summarize.Taxa
Make taxa summary tables similar to QIIME summarize_taxa.py. Summarizes through the first 7 levels of taxonomy of OTUs/SVs wherein the levels are labeled as:
\cr 1-Kingdom
\cr 2-Phylum
\cr 3-Class
\cr 4-Order
\cr 5-Family
\cr 6-Genus
\cr 7-Species
\cr 8- Features with full taxonomy strings}
\usage{
SummarizedTaxa<-Summarize.Taxa(SVtable, TaxTable)
}
\arguments{
\item{FEATURES}{Table of feature/OTU/SV COUNTS where Samples are columns, and IDs are row names}

\item{TAXONOMY}{Table of taxonomies to aggregate on as identified above wherein the column names correspond to the taxonomic level}
}
\value{
A Named List of data frames of each level. Access by like TABLE$Genus or TABLE[[i]] where i is taxonomic level above.
}
\description{
Summarize.Taxa
Make taxa summary tables similar to QIIME summarize_taxa.py. Summarizes through the first 7 levels of taxonomy of OTUs/SVs wherein the levels are labeled as:
\cr 1-Kingdom
\cr 2-Phylum
\cr 3-Class
\cr 4-Order
\cr 5-Family
\cr 6-Genus
\cr 7-Species
\cr 8- Features with full taxonomy strings
}
