% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/The.Works.R
\name{The.Works}
\alias{The.Works}
\title{The.Works}
\usage{
Values<-TheWorks(MicrobeR.Demo.SVtable,MicrobeR.Demo.Tree)
}
\arguments{
\item{FEATURES}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names.}

\item{TREE}{A phylogenetic tree for UniFrac and phylogenetic distances.}

\item{VERBOSE}{Should progress and metrics be printed to screen via message()? Default=TRUE}
}
\value{
A named list containing the following:
\cr.$Features$RawCounts- The original input file of raw counts
\cr.$Features$subsampled- A table of counts sampled to the lowest read count
\cr.$Features$Filtered- Counts where features present in less than 3 samples with less than 10 reads were removed
\cr.$Features$CLR- A CLR-normalized table (see Make.CLR)
\cr.$Features$Percent- The features as percent abundance (sum to 100)
\cr.$Features$Proportion- The feature as straight proportions (sum to 1)
\cr.$Features$PhILR- The Phylogenetic isometric log ratio transform (see package information for philr)
\cr.$Features$Tree$Raw- The input Tree
\cr.$Tree$... The corresponding pruned trees for the feature tables aboeve
\cr.$AlphaDiversity- A number of alpha diversity metrics in a data frame
\cr.$DistanceMatrices- A named list of distance matrices for UniFrac (w/uw), bray curtis, jsd, jaccard, PhILR euclidian, and CLR euclidian
\cr.$PCoA$Vectors- A named list numeric coordinates for plotting PCoA analysis
\cr.$PCoA$Scree- A table containing the percent variation explained for all calculated distance matrices to help in selecting the most appropriate metric.
}
\description{
Takes a feature/OTU/SV table and calculates an extensive set of metrics for downstream use. A number of versions of the normalized data are available from the returned object. See below for more information. For extremely large tables (>500x1000) this function is not reccomended as it will make numerous copies and have a significant memory footprint.
}
