% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyConvert.R
\name{TidyConvert.WhatAmI}
\alias{TidyConvert.WhatAmI}
\title{Convert between table formats as needed (used behind the scenes)}
\usage{
TidyConvert.WhatAmI(TABLE)
}
\arguments{
\item{TABLE}{Table of feature/OTU/SV counts where Samples are columns, and feature IDs are row names.}

\item{KEY}{The column name for a unique identifier.}
}
\value{
Converted table
}
\description{
\code{TidyConvert} Detects formatting of tables and interconverts between data.frame/data.table/matrix/tibble as needed.
When a format is encountered that does not use row names (data.table, tibble), the first column is always expected to be the key although right now a user supplied key is required.
}
