% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make.Percent.R
\name{Make.Percent}
\alias{Make.Percent}
\title{\code{Make.Percent} Takes a table of Features/OTUs/SVs and converts to Percentages (Sum to 100).}
\usage{
Make.Percent(TABLE)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}
}
\value{
Table of Percents
}
\description{
\code{Make.Percent} Takes a table of Features/OTUs/SVs and converts to Percentages (Sum to 100).
}
