% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.Replicates.R
\name{Merge.Replicates}
\alias{Merge.Replicates}
\title{\code{Merge.Replicates} Sum counts together for replicates (ex. pcr/extraction/sequencing run)}
\usage{
Merge.Replicates(OTUTABLE, METADATA, CATEGORY)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{METADATA}{A fraction to use to filter}

\item{CATEGORY}{Minimum number of reads to pass filter as a string, ex. "RepID"}
}
\value{
A list of [[1]] new Feature/OTU/SV table and [[2]] collapsed metadata file
}
\description{
Collapses a table via aggregate FUN=sum based on a provided column in the datatable.
}
