% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.Replicates.R
\name{Merge.Replicates}
\alias{Merge.Replicates}
\title{\code{Merge.Replicates} Sum counts together for replicates (ex. pcr/extraction/sequencing run)}
\usage{
x<-Merge.Replicates(svtable, metadata, "UniqueId")
newtable<-x$Features
newmeta<-x$Metadata
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{METADATA}{Table of metadata with sample name as row name}

\item{CATEGORY}{A category used to dereplicate sequences , ex. "RepID"}
}
\value{
A named list of [[1]] new Feature/OTU/SV table and [[2]] collapsed metadata file
}
\description{
Collapses a table via aggregate FUN=sum based on a provided column in the datatable.
}
