% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microbiome.Barplot.R
\name{Microbiome.Barplot}
\alias{Microbiome.Barplot}
\title{\code{Microbiome.Barplot.R} Create a stacked barplot showing composition of samples,}
\usage{
Microbiome.Barplot(OTUTABLE, METADATA, NTOPLOT, CATEGORY)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{METADATA}{A Table of metadata where sample names are row names.}

\item{NTOPLOT}{A number of features to plot}

\item{CATEGORY}{A Metadata category to block samples by}
}
\value{
Subsampled Table
}
\description{
Uses ggplot2 to create a stacked barplot, for example on phylum level abundances. The most abundant features (defaults to 10, based on rowMeans) will be plotted unless user specified. Anything of over 10 features will use default coloring which will be very difficult to interpret.
}

