% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCoA.R
\name{PCoA}
\alias{PCoA}
\title{\code{PCoA} Generate Distance/Dissimilarities and plot PCoA}
\usage{
PCoA(METRIC="braycurtis", METADATA=experiment_metadata, OTUTABLE=otutable, TREE=ggtree, COLOR="Time", SHAPE="Group", SUBSAMPLE=TRUE, AXIS=c(1,2),)
}
\arguments{
\item{METRIC}{Desired beta-diversity metric, options are Bray Curtis (braycurtis), Weighted UniFrac (weightedunifrac), UnWeighted UniFrac (unweightedunifrac), Jensen-Shannon diversgence (jsd).}

\item{METADATA}{Metadata Table with variables to color PCoAs by}

\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{TREE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names, only required for UniFrac Metrics}

\item{COLOR}{Metadata column to color samples by. Defaults to sample names from metadata row names.}

\item{SHAPE}{Metadata column to shape samples by. Defaults to none.}

\item{SUBSAMPLE}{Should table be subsampled? TRUE/FALSE (default: TRUE)}

\item{AXIS}{Which Axis should be plotted? Expects a numeric vector of length 2. defaults to Pco1 and Pco2 AXIS=c(1,2)}

\item{ADONIS}{Should ADONIS test be applied ? TRUE/FALSE (default TRUE)}
}
\value{
Returns a ggplot
}
\description{
UniFrac is implimented as per GUniFrac, Bray Curtis from Vegan, Jensen-Shannon divergence from textmineR, and PCoA from APE. Plotting via ggplot2. Will also carry out ADONIS and return p and r2 in plot title.
}

