% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.Filter.R
\name{Read.Filter}
\alias{Read.Filter}
\title{\code{Read.Filter} Remove samples with less than a user specified number of reads.}
\usage{
filter<-Read.Filter(svtable, metadata, 5000)
newtable<-filter$Features
newmetadata<-filter$Metadata
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names.}

\item{METADATA}{Metadata file where samples are row names.}

\item{READCUTOFF}{Minimum number of reads to pass filter}

\item{VERBOSE}{Should lists of which samples were removed or kept be printed to screen? (defaults to TRUE)}

\item{PLOT}{Should a plot of read depth be made? (defaults to TRUE)}
}
\value{
A named list containing new feature table and metadata.
}
\description{
\code{Read.Filter} Remove samples with less than a user specified number of reads.
}
