% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.Filter.R
\name{Read.Filter}
\alias{Read.Filter}
\title{\code{Read.Filter} Remove samples with less than a user specified number of reads.}
\usage{
Read.Filter(OTUTABLE, METADATA, READCUTOFF)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{METADATA}{A fraction to use to filter}

\item{READCUTOFF}{Minimum number of reads to pass filter}
}
\value{
A list containing new OTUtable and Metadata. Also creates a histogram of read counts
}
\description{
\code{Read.Filter} Remove samples with less than a user specified number of reads.
}

