% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsample.Table.R
\name{Subsample.Table}
\alias{Subsample.Table}
\title{\code{Subsample.Table} Rarify/Subsample a Feature}
\usage{
Subsample.Table(OTUTABLE, DEPTH, SEED)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{DEPTH}{Count depth, defaults to min(colSums(OTUTABLE)) if not passed}

\item{SEED}{A randomization SEED, defaults to 182 as tribute to scarface and blink.}
}
\value{
Subsampled Table
}
\description{
Takes a Feature/OTU/SV table where samples are columns and feature names are row names. This was formerly using the VEGAN rrarefy function but was found to be slow on large numbers of samples (~3000). It is coded using a loop while the implementation in phyloseq is vectorized it would appear. Now this function is really just an alias for rarefy_even_depth from phyloseq. NOTE: Sampling with replacement!
}

