% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subsample.Table.R
\name{Subsample.Table}
\alias{Subsample.Table}
\title{\code{Subsample.Table} Rarify/Subsample a Feature}
\usage{
Subsample.Table(OTUTABLE, DEPTH, SEED)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV counts where Samples are columns, and IDs are row names}

\item{DEPTH}{Count depth, defaults to min(colSums(OTUTABLE)) if not passed}

\item{SEED}{A randomization SEED, defaults to not specificed for VEGAN.}
}
\value{
Subsampled Table
}
\description{
Takes a Feature/OTU/SV table where samples are columns and feature names are row names, subsamples using VEGAN implementation using randomization seed if provided. Note: samples with replacement
}

