% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize.Taxa.R
\name{Summarize.Taxa}
\alias{Summarize.Taxa}
\title{\code{Summarize.Taxa} Make taxa summary tables similar to QIIME summarize_taxa.py}
\usage{
SummarizedTaxa<-Summarize.Taxa(SVtable, TaxTable)
}
\arguments{
\item{OTUTABLE}{Table of feature/OTU/SV COUNTS where Samples are columns, and IDs are row names}

\item{TAXONOMY}{Table of taxonomies to aggregate on as identified above wherein the column names correspond to the taxonomic level}
}
\value{
A Named List of data frames of each level. Access by like TABLE$Genus or TABLE[[i]] where i is taxonomic level above.
}
\description{
Summarizes through the first 7 levels of taxonomy of OTUs/SVs wherein the levels are labeled as:
1-Kingdom
2-Phylum
3-Class
4-Order
5-Family
6-Genus
7-Species
8- Features with full taxonomy strings
}
